64-bit fixes  [Oliver Stoeneberg]
git-svn-id: svn://messdev.no-ip.org/mess@671 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/machine/spectrum.c
@@ -1026,7 +1026,7 @@ void spectrum_setup_z80(unsigned char *pSnapshot, unsigned long SnapshotSize)
 			/* go to next block */
 			pSource += (3 + length);
 		}
-		while (((unsigned long) pSource - (unsigned long) pSnapshot) < SnapshotSize);
+		while ((pSource - pSnapshot) < SnapshotSize);
 
 		if ((spectrum_128_port_7ffd_data != -1) && (z80_type != SPECTRUM_Z80_SNAPSHOT_48K))
 		{
